+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 ## [0.6.1] 2017-11-13
 ### Added
-- Support for users to use programmatically generated crdentials and cluster config
+- Support for users to use programmatically generated credentials and cluster config
 
 ## [0.6.0] 2017-11-03
 ### Added